pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "otel-logs-collector"
version = "0.1.0"
description = "Transform CDN logs from S3 to OpenTelemetry and send to Grafana Alloy"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
]

[project.scripts]
cdn-logs-to-otel = "src.collector:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
